🐛 fix internal layout: move main.go boilerplate to cmd/ directory (#4246)
fix internal layout: move main.go boilerplate to cmd/ directory

Moves main.go boilerplate to the cmd/ directory to align the project structure with scaffolded templates, ensuring consistency and maintainability.
