+++ pkg/plugins/golang/v4/scaffolds/internal/templates/cmd/main.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package cmd
 
 import (
 	"fmt"
